spaceapp/requirements.txt
asgiref==3.9.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
coreschema==0.0.4
cryptography==46.0.1
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.2
numpy==2.3.3
nvidia-nccl-cu12==2.28.3
pandas==2.3.3
pycparser==2.23
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wheel==0.45.1
xgboost==3.0.5
